% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{f1}
\alias{f1}
\title{GLopR1 package}
\usage{
f1(x)
}
\arguments{
\item{x}{numeric variable}
}
\description{
Given a vector dataset, the GLopR1 package creates its histogram, boxplot,
Q-Q plot, and descriptive statistics summary.
}
\examples{
x=runif(100,1,2) # Creating a vector dataset.
f1(x)            # Apply the function to create histogram, boxplot,
                 # Q-Q plot and descriptive statistics summary of the dataset.
}
\references{
1. Wickham, Hadley (2015). R Packages. O'Reilly Media, Inc. ISBN: 9781491910597.

2. Agresti, Alan. Statistical Methods for the Social Sciences 5th Edition. Pearson, ISBN-13:978-0134507101.
}
